+++ packages/api/lambdas/sf-starter.js
@@ -24,8 +24,6 @@ const {
 function dispatch(message) {
   const input = Object.assign({}, message.Body);
 
-  input.cumulus_meta.workflow_start_time = Date.now();
-
   if (!input.cumulus_meta.execution_name) {
     input.cumulus_meta.execution_name = uuidv4();
   }