+++ CHANGELOG.md
@@ -37,12 +37,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **CUMULUS-1654**
   - Changed `@cumulus/common/SfnStep.parseStepMessage()` to a static class method
 
+- **CUMULUS-1641**
+  - Added `meta.retries` and `meta.visibilityTimeout` properties to sqs-type rule. To create sqs-type rule, you're required to configure a dead-letter queue on your queue.
+  - Added `sqsMessageRemover` lambda which removes the message from SQS queue upon successful workflow execution.
+  - Updated `sqsMessageConsumer` lambda to not delete message from SQS queue, and to retry the SQS message for configured number of times.
+
 ### Removed
 
 - Removed `create_service_linked_role` variable from `tf-modules/data-persistence` module.
 
 ### Fixed
 
+- **CUMULUS-1668** - Fixed a race condition where executions may not have been
+  added to the database correctly
 - **CUMULUS-1654** - Fixed issue with `publishReports` Lambda not including workflow execution error information for failed workflows with a single step
 - Fixed `tf-modules/cumulus` module so that the `urs_url` variable is passed on to its invocation of the `tf-modules/archive` module